requirements.txt
chromadb
google-generativeai
